+++ Common/src/main/java/at/petrak/hexcasting/common/casting/actions/spells/OpBlink.kt
@@ -10,6 +10,7 @@ import at.petrak.hexcasting.api.casting.getEntity
 import at.petrak.hexcasting.api.casting.iota.Iota
 import at.petrak.hexcasting.api.casting.mishaps.MishapBadLocation
 import at.petrak.hexcasting.api.casting.mishaps.MishapImmuneEntity
+import at.petrak.hexcasting.api.casting.toNonZeroLong
 import at.petrak.hexcasting.api.misc.MediaConstants
 import at.petrak.hexcasting.api.mod.HexConfig
 import at.petrak.hexcasting.api.mod.HexTags
@@ -17,7 +18,6 @@ import at.petrak.hexcasting.common.casting.actions.spells.great.OpTeleport
 import net.minecraft.world.entity.Entity
 import net.minecraft.world.phys.Vec3
 import kotlin.math.absoluteValue
-import kotlin.math.roundToLong
 
 object OpBlink : SpellAction {
     override val argc = 2
@@ -54,7 +54,7 @@ object OpBlink : SpellAction {
 
         return SpellAction.Result(
             Spell(target, dvec),
-            (MediaConstants.SHARD_UNIT * delta.absoluteValue * 0.5).roundToLong(),
+            (MediaConstants.SHARD_UNIT * delta.absoluteValue * 0.5).toNonZeroLong(),
             listOf(
                 ParticleSpray.cloud(targetMiddlePos, 2.0, 50),
                 ParticleSpray.burst(targetMiddlePos.add(dvec), 2.0, 100)